fix(ci): resolve storefront static analysis errors
Narrow route-bound models, document cast-backed model attributes, and handle nullable relationships with explicit control flow. Restore soft-deleted wishlist rows through Eloquent and cover re-adding a removed wishlist item.\n\nVerified with XDEBUG_MODE=coverage herd composer ci:check (212 tests, 2079 assertions, 89.2% coverage).

+++ app/Http/Middleware/HandleInertiaRequests.php
@@ -50,7 +50,7 @@ class HandleInertiaRequests extends Middleware
                     ->exists() ?? false,
             ],
             'commerce' => [
-                'cart_quantity' => fn (): int => $request->user()?->cart?->items()->sum('quantity') ?? 0,
+                'cart_quantity' => fn (): int => (int) ($request->user()?->cart?->items()->sum('quantity') ?? 0),
                 'wishlist_count' => fn (): int => $request->user()?->watchlistEntries()->count() ?? 0,
             ],
             'marketplace' => config('marketplace'),
